# PERF: Improved runtime performance