@tillig 2018-05-04
I think this all looks pretty good. Thank you very much for the PR!